// backend/src/dto/systems.dto.ts
export class FsoSystemDto {
  id: number;
  name: string;
  txPower: number;
  txPowerUnit: string;
  rxSensitivity: number;
  rxLensDiameter: number;
  directivity: number;
  wavelength: number;
  companyName: string;
}

export class RfSystemDto {
  id: number;
  name: string;
  txPower: number;
  txPowerUnit: string;
  rxSensitivity: number;
  frequency: number;
  gain: number;
  companyName: string;
}